pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "seldon"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "neo4j>=5.7",
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.scripts]
seldon = "seldon.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["seldon*"]

[tool.setuptools.package-data]
"seldon.domain" = ["*.yaml"]
